recorder-module-4/pyproject.toml
[tool.pdm]
distribution = false

[tool.pdm.build]
includes = []
[project]
name = "recorder-module-1"
version = "0.1.0"
description = ""
authors = [
    {name = "pikamonvvs", email = "pikamonvvs@naver.com"},
]
dependencies = [
    "anyio>=4.4.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.27.0",
    "h2>=4.1.0",
    "streamlink>=6.8.3",
    "httpx-socks>=0.9.1",
    "loguru>=0.7.2",
    "pyinstaller>=6.9.0",
    "types-requests>=2.32.0.20240712",
]
requires-python = "<3.13,>=3.12"
readme = "README.md"
license = {text = "MIT"}

